feat: 設定にデータ管理ページを新設 (#187)
全般ページの肥大化を解消するため、設定ファイルのエクスポートと
保存済み検索クエリを新設の「データ管理」ページに移動。
全般にはテーマ・言語・タイムラインの既定値が残る。
