+++ Views/Settings/GeneralPage.xaml
@@ -2,8 +2,7 @@
     x:Class="XTimelineViewer.Views.Settings.GeneralPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:controls="using:CommunityToolkit.WinUI.Controls"
-    xmlns:local="using:XTimelineViewer.Views.Settings">
+    xmlns:controls="using:CommunityToolkit.WinUI.Controls">
 
     <ScrollViewer Padding="24,16">
         <StackPanel Spacing="4">
@@ -31,14 +30,6 @@
                           SelectionChanged="LanguageCombo_SelectionChanged"/>
             </controls:SettingsCard>
 
-            <!-- Export Settings Folder -->
-            <controls:SettingsCard x:Name="ExportFolderCard">
-                <controls:SettingsCard.HeaderIcon>
-                    <FontIcon Glyph="&#xE838;" FontFamily="Segoe Fluent Icons"/>
-                </controls:SettingsCard.HeaderIcon>
-                <Button x:Name="OpenFolderBtn" Click="OpenFolderBtn_Click"/>
-            </controls:SettingsCard>
-
             <!-- Timeline Defaults -->
             <controls:SettingsExpander x:Name="DefaultTimelineExpander">
                 <controls:SettingsExpander.HeaderIcon>
@@ -61,24 +52,6 @@
                                       MinWidth="0"/>
                     </controls:SettingsCard>
                 </controls:SettingsExpander.Items>
-            </controls:SettingsExpander>
-
-            <!-- Saved Search Queries -->
-            <controls:SettingsExpander x:Name="SavedQueriesExpander"
-                                       ItemsSource="{x:Bind SavedQueries}">
-                <controls:SettingsExpander.HeaderIcon>
-                    <FontIcon Glyph="&#xE721;" FontFamily="Segoe Fluent Icons"/>
-                </controls:SettingsExpander.HeaderIcon>
-                <controls:SettingsExpander.ItemTemplate>
-                    <DataTemplate x:DataType="local:SavedQueryItem">
-                        <controls:SettingsCard Header="{x:Bind Decoded}">
-                            <Button Content="{x:Bind DeleteLabel}"
-                                    Tag="{x:Bind Path}"
-                                    Click="SavedQueryDelete_Click"
-                                    Foreground="{ThemeResource SystemFillColorCriticalBrush}"/>
-                        </controls:SettingsCard>
-                    </DataTemplate>
-                </controls:SettingsExpander.ItemTemplate>
             </controls:SettingsExpander>
         </StackPanel>
     </ScrollViewer>
