+++ Views/Settings/GeneralPage.xaml.cs
@@ -3,25 +3,11 @@ using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Navigation;
 using System;
 using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.IO;
-using System.Linq;
 
 namespace XTimelineViewer.Views.Settings
 {
-    /// <summary>保存済み検索クエリの表示用アイテム。x:Bind から参照される。</summary>
-    public sealed class SavedQueryItem(string path, string decoded, string deleteLabel)
-    {
-        public string Path        { get; set; } = path;
-        public string Decoded     { get; set; } = decoded;
-        public string DeleteLabel { get; set; } = deleteLabel;
-    }
-
     public sealed partial class GeneralPage : Page
     {
-        /// <summary>x:Bind のバインディングソース。XAML から参照される。</summary>
-        public ObservableCollection<SavedQueryItem> SavedQueries { get; } = [];
-
         private static readonly string[] ThemeValues = ["Default", "Light", "Dark"];
         private static readonly string[] LangValues  = ["system", "ja-JP", "en-US"];
 
@@ -74,11 +60,6 @@ namespace XTimelineViewer.Views.Settings
             var langIdx = Array.IndexOf(LangValues, _parent?.Settings.Language ?? "system");
             LanguageCombo.SelectedIndex = langIdx < 0 ? 0 : langIdx;
 
-            // Export Folder
-            ExportFolderCard.Header      = R.Get("Settings_ExportFolder");
-            ExportFolderCard.Description = _parent?.SettingsFolder ?? "";
-            OpenFolderBtn.Content        = R.Get("Button_OpenFolder");
-
             // Timeline Defaults
             DefaultTimelineExpander.Header      = R.Get("Settings_DefaultTimeline");
             DefaultTimelineExpander.Description = R.Get("Settings_DefaultTimeline_Description");
@@ -102,48 +83,7 @@ namespace XTimelineViewer.Views.Settings
                 ListHeaderToggle.IsOn       = !s.DefaultHideListHeader;
             }
 
-            // Saved Search Queries
-            SavedQueriesExpander.Header      = R.Get("Settings_SavedQueries");
-            SavedQueriesExpander.Description = R.Get("Settings_SavedQueries_Description");
-            RebuildSavedQueriesItems();
-
             _isInitializing = false;
-        }
-
-        private void RebuildSavedQueriesItems()
-        {
-            SavedQueries.Clear();
-            if (_parent is null) return;
-
-            var deleteLabel = R.Get("Profile_Delete");
-            foreach (var path in _parent.Settings.SavedSearchQueries)
-                SavedQueries.Add(new SavedQueryItem(path, Uri.UnescapeDataString(path), deleteLabel));
-
-            UpdateSavedQueriesExpanderState();
-        }
-
-        private void UpdateSavedQueriesExpanderState()
-        {
-            if (SavedQueries.Count == 0)
-            {
-                SavedQueriesExpander.IsExpanded = false;
-                SavedQueriesExpander.IsEnabled  = false;
-            }
-            else
-            {
-                SavedQueriesExpander.IsEnabled = true;
-            }
-        }
-
-        private void SavedQueryDelete_Click(object sender, RoutedEventArgs e)
-        {
-            if (_parent is null || sender is not Button btn || btn.Tag is not string path) return;
-            _parent.Settings.SavedSearchQueries.Remove(path);
-            _parent.NotifySettingsChanged();
-
-            var item = SavedQueries.FirstOrDefault(q => q.Path == path);
-            if (item is not null) SavedQueries.Remove(item);
-            UpdateSavedQueriesExpanderState();
         }
 
         private void ThemeCombo_SelectionChanged(object sender, SelectionChangedEventArgs e)
@@ -171,14 +111,6 @@ namespace XTimelineViewer.Views.Settings
             var idx = Math.Clamp(LanguageCombo.SelectedIndex, 0, LangValues.Length - 1);
             _parent.Settings.Language = LangValues[idx];
             _parent.NotifySettingsChanged();
-        }
-
-        private async void OpenFolderBtn_Click(object sender, RoutedEventArgs e)
-        {
-            if (_parent is null) return;
-            var folder = _parent.SettingsFolder;
-            Directory.CreateDirectory(folder);
-            await Windows.System.Launcher.LaunchFolderPathAsync(folder);
         }
 
         private void SidebarToggle_Toggled(object sender, RoutedEventArgs e)
